# optimize refine perm for PooledArray columns by sorting pool once