--- a/src/StructArrays.jl
+++ b/src/StructArrays.jl
@@ -15,6 +15,10 @@ function __init__()
     Requires.@require Tables="bd369af6-aec1-5ad0-b16a-f7cc5008161c" include("tables.jl")
     Requires.@require PooledArrays="2dfb63ee-cc39-5dd5-95bd-886bf059d720" begin
         fastpermute!(v::PooledArrays.PooledArray, p::AbstractVector) = permute!(v, p)
+        function sort_by(y::PooledArrays.PooledArray)
+            poolranks = invperm(sortperm(y.pool))
+            j->(@inbounds k=poolranks[y.refs[j]]; k)
+        end
     end
     Requires.@require WeakRefStrings="ea10d353-3f73-51f8-a26c-33c1cb351aa5" begin
         fastpermute!(v::WeakRefStrings.StringArray, p::AbstractVector) = permute!(v, p)