--- a/src/sort.jl
+++ b/src/sort.jl
@@ -64,11 +64,12 @@ end
 Base.sort!(c::StructArray{<:Union{Tuple, NamedTuple}}) = permute!(c, sortperm(c))
 Base.sort(c::StructArray{<:Union{Tuple, NamedTuple}}) = c[sortperm(c)]
 
+sort_by(y) = j->(@inbounds k=y[j]; k)
 # Methods from IndexedTables to refine sorting:
 # # assuming x[p] is sorted, sort by remaining columns where x[p] is constant
 function refine_perm!(p, cols, c, x, y, lo, hi)
     temp = similar(p, 0)
-    order = Base.Order.By(j->(@inbounds k=y[j]; k))
+    order = Base.Order.By(sort_by(y))
     nc = length(cols)
     for (_, idxs) in TiedIndices(x, p, lo:hi)
         i, i1 = extrema(idxs)