So far we were using the optimized sortperm only when the first column was a PooledArray. This adds some optimizations if any other column is a PooledArray

Before
```
julia> a=rand(1:100, 10^6); b=PooledArray(rand([randstring(10) for i=1:10], 10^6));
c = StructVector((b=b, a=a))
julia> @btime sortperm(c);
  29.188 ms (47 allocations: 9.94 MiB)
```
```
c = StructVector((a=a, b=b))
julia> @btime sortperm(c);
  172.249 ms (8 allocations: 7.74 MiB)
```

After:
```julia
# speed when first column is a PooledArray remains the same
c = StructVector((b=b, a=a))
julia> @btime sortperm(c);
  28.748 ms (47 allocations: 9.93 MiB)
```

```
# this improves 3x since we don't do stringcmp at all.
c = StructVector((a=a, b=b))
julia> @btime sortperm(c);
  54.885 ms (11 allocations: 7.75 MiB)
```