@shashi 2019-02-09
I guess we need to add code to handle previous versions of PooledArrays where `y.pool` was a Dict.

I'm thinking of adding an `isa` branch