@piever 2019-02-10
Nice to see this performance improvement is possible!

I'm not sure it's a good idea to put this kind of code under a `@require` block: it seems like it relies to much on the internal implementation of PooledArrays and I cannot control which version the user has without taking a dependency.

I see three possible ways forward:

1) Either StructArrays takes a dependency on PooledArrays or viceversa. Having StructArrays depend on PooledArrays feels like a design that doesn't scale: it's strange to add a dependency to all abstract array packages that have optimized implementations of sorting.

2) Make a tiny SortingBase package where the `sort_by` and `fastpermute!` fallbacks would live, then WeakRefStrings and PooledArrays would extend things from there. This may be the cleanest solution: a tiny dependency-free package to define a more powerful interface for sorting.

3) Ideally SortingBase is not needed and one can do the same by overloading some base functions. Instead of `Base.Order.By(k -> y[k])` one could use `Base.Order.Perm(Base.Order.Forward, y)` and overload it for pooled arrays. Unfortunately I'm not 100% sure this would be OK as you couldn't actually return a `Perm`: I wonder whether that's OK as long as it respects the `Ordering` interface. It may be worth opening an issue in Base to discuss what's the proper way for an array type to optimize `Perm`.