@shashi 2019-02-11
I like option 3 the most.