@shashi 2019-02-11
I changed this to have a `fast_sortable` that returns a version of the array (in the PooledArray case just a PooledArray of Integers) which leads to the counting sort subroutine being chosen for refine_perm. This leads to a further 2x improvement in performance!

```
julia> @btime sortperm(c);
  22.106 ms (123 allocations: 7.88 MiB)
```