@shashi 2019-02-11
BTW there's also https://github.com/JuliaCollections/SortingAlgorithms.jl where some of this abstraction can potentially live (as in option 2)