@shashi 2019-02-12
I think I'm having to create interactions between PooledArrays and StringArrays for sort performance. I guess src/sort.jl itself a candidate for a separate package which can depend on PooledArrays and WeakRefStrings.