@piever 2019-02-12
I'm actually happy to take some extra dependency here if I can move out all this sorting code: it's a bit weird that such a large chunk of StructArrays code is about sorting optimizations.

I'm not sure whether this other package should be SortingAlgorithms or a new one, but feel free to move things from `src/sort.jl` (plus `fastpermute!` and `fast_sortable`) to another place: happy to take a look if/when that happens!