@shashi 2019-02-12
> such a large chunk of StructArrays code is about sorting optimizations.

Well this is the most important function as far as performance of IndexedTables is concerned, and performance improvements tend to create a lot of code. :-)

If we do refactor sorting into a separate package will you endorse some type piracy? Let's say StructArraySorting.jl defines `sortperm(::StructArray)`.

Do you have uses for this package other than in IndexedTables? It only makes sense to do this kind of refactoring if we're aiming to have something minimal for certain purposes. Otherwise I would just have this package depend on the other two Array packages and implement the optimizations here first-class. An excellent reason to do this refactoring would be if DataFrames ecosystem wants to use this package as sort of the common currency. cc @nalimilan