@piever 2019-02-12
> Do you have uses for this package other than in IndexedTables? It only makes sense to do this kind of refactoring if we're aiming to have something minimal for certain purposes.

Some people use it on the GPU, see https://github.com/piever/StructArrays.jl/issues/45, or for arrays of custom structs, but I think the table ecosystem has converged to Tables as a common interface.

> Otherwise I would just have this package depend on the other two Array packages and implement the optimizations here first-class.

PooledArrays and WeakRefStrings are rather minimal anyway and it does make sense to have them as otherwise StructArrays would be very inefficient for non-isbits types, so I'm not opposed to adding them if that's helpful. However I'd rather avoid special-casing them too much (or using too much of their internals here) and would prefer to set up a tiny package (potentially in a shared org like JuliaData) with some methods that PooledArrays, WeakRefStrings and any other new array package could overload if they have implementations that are faster than the fallback.