@shashi 2019-02-12
> ecosystem has converged to Tables as a common interface.

Ok then

I think making the right abstractions for a 0-dependency package that allows optimizing array types for sorting when used in StructArrays is a non-trivial amount of work.