@piever 2019-02-13
> I think making the right abstractions for a 0-dependency package that allows optimizing array types for sorting when used in StructArrays is a non-trivial amount of work.

I see, then we can leave that plan for the future (when that interface becomes clearer). If it's just `fastpermute!`, `fast_sortable` (and maybe restoring `compact_mem`) I agree they can just live here. Could you replace the `@require` block for pooled arrays with a dependency in this PR? We may add 0.5 as a lower bound so we can use the simpler code.