@KristofferC 2019-02-13
> Do you have uses for this package other than in IndexedTables?

I think there is a lot of use for a simple package that pretty much only does the SoA transformation well. Personally, I would be a bit sad to see this package become too "bloated" by different data analysis stuff.