@piever 2019-02-13
The whole idea behind this package is to be the backbone for both simple SoA transformation *and* for data tables (seen as a "named tuple" SoA). I'm slowly starting to understand what could be the issue here: IndexedTables, to make use of this package, needs a super optimized method for `sortperm` on a `StructArray`, with the tabular use case in mind: where columns are often string arrays with few unique entries. For that to exist without type piracy, it needs to stay here. On the other hand, I'm reluctant to pool columns by default while doing `sortperm` or change the `eltype` of things: I'm afraid that outside of the "table" use-case, this is not even an optimization (if all entries of the vector are different) and could break things for other array types (say GPU based). One counter-proposal (briefly discussed with @shashi on slack) is the following:
- remove the optimized `sortperm` from here: the fallback `sortperm` works just fine and gives the correct result
- remove all dependency from PooledArrays and WeakRefStrings from here (even the `@require` blocks)
- move all the sorting optimizations to IndexedTables (where they are actually needed) but call this efficient method `fastortperm` (which would use the various `fastpermute!`, `fastsort` and so on, which are sorting things optimized for the case of string arrays with few unique entries) to avoid type piracy and to make sure that the default `sortperm` works on all array types